add function to simulate translate and draw line

+++ Chapter02/Example_2_5.cpp
@@ -23,7 +23,7 @@ bool Example_2_5::UserRender(int elapsed_time)
             auto dragForce = liquid_->CalculateDrag(mover_[i]);
             mover_[i]->ApplyForce(dragForce);
         }
-        PVector gravity(0, 0.1f);
+        PVector gravity(0.0f, 0.1f);
         gravity.Mult((float)mover_[i]->GetMass());
         mover_[i]->ApplyForce(gravity);
         mover_[i]->Update();
@@ -40,7 +40,7 @@ void Example_2_5::Reset(void)
         if (mover_[i] != nullptr) {
             delete mover_[i];
         }
-        mover_[i] = new Mover_2_5(PVector((float)(30 + WindowWidth() / 10 * i), 0), (int)((rand() % 3) + 1.0f));
+        mover_[i] = new Mover_2_5(PVector(30 + WindowWidth() / 10 * i, 0), (int)((rand() % 3) + 1.0f));
     }
 }
